+++ permissions/groups/import.ps1
@@ -1,0 +1,54 @@
+####################################################################
+# HelloID-Conn-Prov-Target-Myneva-ImportPermissions-Group
+# PowerShell V2
+####################################################################
+
+# Enable TLS1.2
+[System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor [System.Net.SecurityProtocolType]::Tls12
+
+#region functions
+#endregion
+
+try {
+    Write-Information 'Initializing Myneva configuration'
+    Import-Module "$($actionContext.Configuration.PowerShellModulePath)" -Force
+
+    $splatMynevaSession = @{
+        Username         = $actionContext.Configuration.UserName
+        Password         = ConvertTo-SecureString -String "$($actionContext.Configuration.Password)"  -AsPlainText -Force
+        WebServiceUrl    = $actionContext.Configuration.BaseUrl
+        FilePathDLL      = $actionContext.Configuration.FilePathDLL
+        OrganisationCode = $actionContext.Configuration.OrganizationCode
+    }
+    $null = New-RegasSession @splatMynevaSession
+
+
+    Write-Information 'Starting Myneva permission entitlement import'
+    $importedPermissions = (Get-RegasGroupList).GroupAssignments
+
+    foreach ($importedPermission in $importedPermissions) {
+        $permission = @{
+            PermissionReference = @{
+                Reference = $importedPermission.group_id
+            }
+            Description         = "$($importedPermission.group_id)"
+            DisplayName         = "$($importedPermission.group_id)"
+            AccountReferences   = @($importedPermission.person_id)
+        }
+        Write-Output $permission
+    }
+    Write-Information 'Myneva permission entitlement import completed'
+}
+catch {
+    $ex = $PSItem
+    if ($($ex.Exception.GetType().FullName -eq 'Microsoft.PowerShell.Commands.HttpResponseException') -or
+        $($ex.Exception.GetType().FullName -eq 'System.Net.WebException')) {
+        $errorObj = Resolve-MynevaError -ErrorObject $ex
+        Write-Warning "Error at Line '$($errorObj.ScriptLineNumber)': $($errorObj.Line). Error: $($errorObj.ErrorDetails)"
+        Write-Error "Could not import Myneva permission entitlements. Error: $($errorObj.FriendlyMessage)"
+    }
+    else {
+        Write-Warning "Error at Line '$($ex.InvocationInfo.ScriptLineNumber)': $($ex.InvocationInfo.Line). Error: $($ex.Exception.Message)"
+        Write-Error "Could not import Myneva permission entitlements. Error: $($ex.Exception.Message)"
+    }
+}