+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [2.0.0] - 11-06-2026
+
+> Since the displayName is removed from the permissions reference This is a breaking change and causes HelloID to lose the reference and see it as a new permission.
+
+### Added
+- Added import and reconciliation functionality
+  - Accounts import Script
+  - Group permissions import Script
+- Added configuration options to import Dynamic Buttons and Services for accounts
+
+
+### Removed
+- Forum link from Readme
+- **Removed DisplayName from the Permissions Reference**
+
 ## [1.0.0] - 29-05-2024
 
 This is the first official release of _HelloID-Conn-Prov-Target-Myneva_. This release is based on template version _1.2.0_.